collector: move internal/collector → pkg/collector
Make the collector package importable outside the module. Go's internal/
rule blocks external import, which #51 needs: external programs embed the
collector in-process, and it's the foundation for the headless gRPC stream.
Its emitted types thus become a public API surface.

Pure move + import-path rewrite — no code changes beyond the path. Only
internal/tui imported it (now points at pkg/collector). The package still
imports internal/bpf, which is allowed within the same module. CLAUDE.md's
structure tree and prose updated to match.

Part of #51 (PR2 of the phased rollout). The protobuf schema + gRPC
bindings land in a follow-up.
