+++ internal/tui/view_timeline.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/charmbracelet/lipgloss"
 
-	"github.com/trentas/ptop/internal/collector"
+	"github.com/trentas/ptop/pkg/collector"
 )
 
 // renderTimelineView (F7) — assets/mockup.jsx → TimelineView
