+++ pkg/collector/clktck_test.go
