+++ pkg/collector/cpu_darwin.go
