+++ pkg/collector/cpu_darwin_test.go
