+++ pkg/collector/cpu_ebpf_stub.go
