+++ pkg/collector/io_darwin.go
