+++ pkg/collector/io_ebpf_stub.go
