+++ pkg/collector/libproc_darwin.go
