+++ pkg/collector/mem_ebpf_stub.go
