+++ pkg/collector/network_darwin_test.go
