+++ pkg/collector/network_ebpf.go
