+++ pkg/collector/network_ebpf_stub.go
