+++ pkg/collector/set.go
