+++ pkg/collector/set_test.go
