+++ pkg/collector/sockets.go
