+++ pkg/collector/sockets_test.go
