+++ pkg/collector/syscalls_ebpf.go
