+++ pkg/collector/syscalls_ebpf_stub.go
