+++ pkg/collector/threads_darwin.go
