+++ pkg/collector/threads_ebpf.go
