+++ pkg/collector/threads_ebpf_stub.go
