+++ pkg/collector/threads_proc.go
